// document-builder/src/simpleDocumentBuilder/types/SimpleDocumentData.ts
export interface SimpleDocumentData {
    family_name: string;
    given_name: string;
    portrait: string;
    birth_date: string;
    issue_date: string;
    expiry_date: string;
    issuing_country: string;
    document_number: string;
    type_of_fish: string;
    number_of_fishing_rods: string;
}